Reorganize `IndexLogic`
