+++ creusot-contracts/src/logic.rs
@@ -8,7 +8,7 @@ mod fmap;
 mod fset;
 mod int;
 mod mapping;
-mod ops;
+pub mod ops;
 pub mod ord;
 mod seq;
 mod set;
@@ -17,7 +17,6 @@ pub use fmap::FMap;
 pub use fset::FSet;
 pub use int::Int;
 pub use mapping::Mapping;
-pub use ops::IndexLogic;
 pub use ord::OrdLogic;
 pub use seq::Seq;
 pub use set::Set;
